requirements.txt
# Core
Flask==2.3.3
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
numpy==1.26.4
matplotlib==3.9.2
scikit-learn
faiss-cpu==1.8.0

# Chỉ dùng CPU (không cần CUDA)
--extra-index-url https://download.pytorch.org/whl/cpu